+++ Assets/Scripts/IBuilding.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IBuilding
 {
-    Dictionary<ResourceType, float> ResourcesUses { get; }
-    Dictionary<ResourceType, float> ResourcesProduces { get; }
+    Dictionary<ResourceType, int> ResourcesUses { get; }
+    Dictionary<ResourceType, int> ResourcesProduces { get; }
     void Set();
 }
